Core/Misc: Move fmt::formatter forward declaration to a common header to reduce number of replacements needed when upgrading fmt to new major version
