+++ src/server/game/Entities/Object/ObjectGuid.h
@@ -20,13 +20,13 @@
 
 #include "Define.h"
 #include "EnumFlag.h"
+#include "StringFormatFwd.h"
 #include "advstd.h"
 #include <array>
 #include <functional>
 #include <list>
 #include <set>
 #include <span>
-#include <stdexcept>
 #include <string>
 #include <type_traits>
 #include <unordered_set>
@@ -406,34 +406,12 @@ struct std::hash<ObjectGuid>
     }
 };
 
-namespace fmt
-{
-inline namespace v10
-{
-template <typename T, typename Char, typename Enable>
-struct formatter;
-
 template <>
-struct formatter<ObjectGuid, char, void>
+struct fmt::formatter<ObjectGuid, char, void> : Trinity::NoArgFormatterBase
 {
-    template <typename ParseContext>
-    constexpr auto parse(ParseContext& ctx) -> decltype(ctx.begin())
-    {
-        auto begin = ctx.begin(), end = ctx.end();
-        if (begin == end)
-            return begin;
-
-        if (*begin != '}')
-            throw std::invalid_argument("invalid type specifier");
-
-        return begin;
-    }
-
     template <typename FormatContext>
     auto format(ObjectGuid const& guid, FormatContext& ctx) const -> decltype(ctx.out());
 };
-}
-}
 
 namespace Trinity
 {
